requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
pydantic==2.9.2
pydantic-settings==2.5.2
apscheduler==3.10.4
httpx==0.27.2
aiosmtplib==3.0.2
python-multipart==0.0.12
itsdangerous==2.2.0
tzdata>=2024.2
pytest-asyncio>=0.24
